backend-main/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
slowapi>=0.1.9
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
faker>=20.0.0
pydantic>=2.0.0
crewai>=0.1.0
litellm>=1.0.0
apscheduler>=3.10.0
